Onboard EOLib.Graphics to using AutomaticTypeMapper
