+++ EOLib.Graphics/IGraphicsDeviceRepository.cs
@@ -2,6 +2,7 @@
 // This file is subject to the GPL v2 License
 // For additional details, see the LICENSE file
 
+using AutomaticTypeMapper;
 using Microsoft.Xna.Framework.Graphics;
 
 namespace EOLib.Graphics
@@ -16,6 +17,8 @@ namespace EOLib.Graphics
         GraphicsDevice GraphicsDevice { get; }
     }
 
+    [MappedType(BaseType = typeof(IGraphicsDeviceRepository), IsSingleton = true)]
+    [MappedType(BaseType = typeof(IGraphicsDeviceProvider), IsSingleton = true)]
     public class GraphicsDeviceRepository : IGraphicsDeviceRepository, IGraphicsDeviceProvider
     {
         public GraphicsDevice GraphicsDevice { get; set; }
