+++ EOLib.Graphics/packages.config
@@ -1,5 +1,6 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="AutomaticTypeMapper" version="1.1.0.2" targetFramework="net462" />
   <package id="MonoGame.Framework.DesktopGL" version="3.7.1.189" targetFramework="net462" />
   <package id="PELoaderLib" version="1.1.0.3" targetFramework="net462" />
   <package id="System.Runtime.CompilerServices.Unsafe" version="4.5.2" targetFramework="net462" />
