+++ EndlessClient/packages.config
@@ -3,6 +3,7 @@
   <package id="AutomaticTypeMapper" version="1.1.0.2" targetFramework="net462" />
   <package id="EndlessClient.Binaries" version="1.0.0" targetFramework="net462" />
   <package id="MonoGame.Framework.DesktopGL" version="3.7.1.189" targetFramework="net462" />
+  <package id="PELoaderLib" version="1.1.0.3" targetFramework="net462" />
   <package id="System.Runtime.CompilerServices.Unsafe" version="4.5.2" targetFramework="net462" />
   <package id="Unity" version="5.9.6" targetFramework="net462" />
   <package id="XNAControls" version="1.1.1" targetFramework="net462" />
